Add NSMutableArray+Manipulation.m/h. Refine coding style.

+++ motcha/Category/UIImage+Scale.h
@@ -1,6 +1,6 @@
 #import <UIKit/UIKit.h>
 
 @interface UIImage (Scale)
-- (UIImage *)resizeToWidth:(float)width height:(float)height;
-- (UIImage *)scaleToMaxWidth:(float) maxWidth maxHeight:(float) maxHeight;
+- (UIImage *)resizeToWidth:(CGFloat)width height:(CGFloat)height;
+- (UIImage *)scaleToMaxWidth:(CGFloat)maxWidth maxHeight:(CGFloat)maxHeight;
 @end
